Change application structure; use go-kit's log
In order to separate the main server binary from the application files,
the binary is moved to the cmd folder. This commit also introduces
go-kit (to be used in the development of this microservice), and starts
making use of its Log.
The makefile changes reflect the above, along with adding the .PHONY
targets (in case of any file conflicts).