Add go-kit REST API functionality before implementation of business logic
This commit makes use of go-kit's features to add a REST interface to
this service. It adds several features such as using the mux library for
request routing, it implements the basic structure for storing encrypted
data in-memory, and it performs the encoding/decoding of requests and
responses.

The commit also adds more information to the README, such as links about
go-kit, how to build the project, and example usage.