Add encryption of stored data
This commit adds the required functionality of storing encrypted data.
This is done by generating a random 32-byte AES key per item, along with
an IV. The procedure is based on examples on
https://golang.org/pkg/crypto/cipher/#example_NewCFBEncrypter and
https://golang.org/pkg/crypto/cipher/#example_NewCFBDecrypter.
The underlying data store has been changed from a map containing as
values byte arrays to strings instead. A bug was occuring wherein
repeated accesses to the same key on the map resulted in different
arrays being returned. As strings in go hold arbitrary bytes and I could
not find the cause of the bug, I decided to change the representation.

The encryption key generated is returned in base64 format, so it will
need to be URL encoded on the client when performing data retrievals.
The README has been updated to reflect this.