Switch back to byte array as underlying storage
This commit fixes an issue found in a previous commit in which data
would be incorrectly presented. The issue manifested itself by on
consecutive calls to retrieving data, correctly and then immediately
incorrectly decrypting the data. For instance, on a first call to
retrieving a given ID, the service would return the correct data, but on
the next call would return an error. Then it would return the correct
data again, and then an error, and so on. In order to debug this, print
statements were added to the result of the retrieval calls (in the
service logic). It was found that the very last few bytes of the byte
arrays were changing, where the first ones were the same. Due to the
storage being byte arrays (compared to strings, which are immutable),
the suspicious was that something was mutable. The next step was then
to add more print statements, but this time on the store endpoint. It
was found that the problem relied on the decrypt function, and it was
immediately found that the XOR part of the function was the issue (or
rather, the fact that it took a splice). To fix this, this commit makes
a copy of said slice onto a different one, thereby not modifying the
original.