client: Implement retrieve call; remove main.go
This commit introduces the implementation of the retrieve call, along
with returning go errors (decoded from returned json) to the client's
caller. It also removes main.go: the rationale is that this client, in
the scope of this repository, will be used in tests. As such having an
example command line interface to the client is not needed. The README
has been updated to include an example client though (and the formatting
has been changed in the make example to make use of code blocks).