Add a few tests
This commit adds a few integration tests (that perform calls using the
client and check for the returned value). In order to do this, a little
refactoring needed to be performed, such as moving the server's
initialisation to a different file on the server package, such that the
tests could start the server.
The client also returns pre-defined errors (instead of always creating
them on the fly), so that the tests can check against them.