server: Move storage onto separate component
This should make it easier to split out the storage onto a separate
implementation (such as another microservice) if needed. It also
alleviates the locking, in that previously the lock would be maintained
during even when decrypting/encrypting. In the new implementation, it is
possible that encryption of data is done needlessly (e.g. when an item
is already present for an ID), but this was found to be acceptable
(given that it should not happen often) when compared to having the
encryption call in the critical section.